Deleted unnecessary pretrained models and modified folder structure.
