--- hubconf.py
@@ -1,10 +1,0 @@
-#hubconf.py
-
-dependencies = ['torch']
-
-import resnet20
-import resnet32
-import resnet44
-import resnet56
-import resnet110
-import resnet1202
