+++ src/hubconf.py
@@ -1,0 +1,5 @@
+#hubconf.py
+
+dependencies = ['torch']
+
+import resnet20
